requirements.txt
aniso8601==7.0.0
asgiref==3.5.0
boto3==1.24.17
botocore==1.27.17
Django==4.0.4
django-ckeditor==6.4.0
django-cors-middleware==1.5.0
django-js-asset==2.0.0
django-storages==1.12.3
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
jmespath==1.0.1
mpmath==1.2.1
numpy==1.22.3
Pillow==9.1.0
promise==2.3
python-dateutil==2.8.2
Rx==1.6.1
s3transfer==0.6.0
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
sympy==1.10.1
text-unidecode==1.3
urllib3==1.26.9
